rename getGridRowCells to getGridCells
